Update tamarin version and refactor dpad controls for legacy runtime support
The version of the 'tamarin' dependency was updated from 2.1.0 to 2.1.3 in the build.gradle file for the project. Significant refactor of dpad controls was also done in 'Main.java' and 'MovingPlayerExampleState.java'. Instead of separate actions for teleporting, turning left and right, and resetting position, all these actions have been combined into a single Dpad action. This was an essential move as a workaround to support the use of the application in older runtimes that may not support the XR_EXT_dpad_binding extension. Additionally, this reduces redundancy.